Split API views and aggregator handler views into separate modules for sanity
